+cookie +logout

+++ src/visits/visits.service.ts
@@ -1,4 +1,4 @@
-import { Injectable } from "@nestjs/common";
+import { HttpException, HttpStatus, Injectable } from "@nestjs/common";
 import { InjectModel } from "@nestjs/mongoose";
 import { Model } from "mongoose";
 import { CreateVisitDto } from "./dto/create-visit.dto";
@@ -6,43 +6,79 @@ import { UpdateVisitDto } from "./dto/update-visit.dto";
 import { Visit, VisitDocument } from "./shemas/visits.schemas";
 
 
+
 @Injectable()
 export class VisitsService {
-    findallwhere(): Promise<Visit[]> {
-        throw new Error('Method not implemented.');
+
+
+    constructor(
+        @InjectModel(Visit.name) private visitModel: Model<VisitDocument>,
+        // private visitService: VisitsService
+    ) {
+    }
+
+    async findById(id: string): Promise<VisitDocument> {
+        return await this.visitModel.findOne({ id }).exec()
     }
 
 
-    constructor(@InjectModel(Visit.name) private visitModel: Model<VisitDocument>) {
-    }
 
-
-    private visits = []
-    
     async getAll(): Promise<Visit[]> {
-        return this.visitModel.find(
-
-        ).exec()
+        return await this.visitModel.find().exec()
     }
 
 
     async getById(id: string): Promise<Visit> {
-        return this.visitModel.findById(id)
+
+
+        try {
+            const isVisitExist = await this.visitModel.findById(id).exec()
+
+            if (!isVisitExist) {
+                throw new HttpException('Not Found', HttpStatus.NOT_FOUND);
+            }
+            return await this.visitModel.findById(id)
+        } catch (error) {
+            throw new HttpException('Not Found', HttpStatus.NOT_FOUND);
+        }
+
+
     }
 
 
     async create(visitdto: CreateVisitDto): Promise<Visit> {
         const newVisit = new this.visitModel(visitdto)
-        return newVisit.save()
+        return await newVisit.save()
     }
 
 
     async remove(id: string): Promise<Visit> {
-        return this.visitModel.findByIdAndRemove(id)
-    }
 
+        try {
+            const isVisitExist = await this.visitModel.findById(id).exec()
+
+            if (!isVisitExist) {
+                return await this.visitModel.findByIdAndRemove(id)
+            }
+        } catch (error) {
+            throw new HttpException('Not Found', HttpStatus.NOT_FOUND);
+
+        }
+    }
 
     async update(id: string, visitdto: UpdateVisitDto): Promise<Visit> {
-        return this.visitModel.findByIdAndUpdate(id, visitdto, { new: true })
+        try {
+            const isVisitExist = await this.visitModel.findById(id).exec()
+
+            if (!isVisitExist) {
+                return await this.visitModel.findByIdAndUpdate(id, visitdto, { new: true })
+            }
+        } catch (error) {
+            throw new HttpException('Not Found', HttpStatus.NOT_FOUND);
+
+        }
+        
     }
 }
+
+
